Manage Connection from Ruby
This removes the lazy-static macro from Rust and moves the
responsibility of initiating the connection from Ruby.  I
wanted to set it up this way because it feels more predictable
to be able to handle connection errors this way.  It also
lights a path to how a system might work with different db
connections a the same time or potentially doing things inside
of a transaction.