Simple Example of Fetching a Collection
A fair number of reads will involve needing to fetch multiple
rows from a database at a time.  This demonistrates an example
of how this might be accomplished.  I realize that it's doing
a fair amount of heap churn... that will be something I want
to refine as I get a better since for what needs to happen.